Place method to get search results in its component, fixed #34
